+++ src/form-elements/FormElementAPINSWLiquorLicence.tsx
@@ -140,6 +140,7 @@ function FormElementAPINSWLiquorLicence({
           isDirty={isDirty}
           setIsDirty={setIsDirty}
           aria-describedby={ariaDescribedby}
+          autoComplete={element.autocompleteAttributes?.join(' ')}
         />
       </FormElementLabelContainer>