+++ src/form-elements/FormElementBarcodeScanner.tsx
@@ -139,6 +139,7 @@ function FormElementBarcodeScanner({
                   disabled={element.readOnly}
                   onBlur={setIsDirty}
                   aria-describedby={ariaDescribedby}
+                  autoComplete={element.autocompleteAttributes?.join(' ')}
                 />
                 <span className="ob-input-icon icon is-small is-right">
                   <MaterialIcon className="is-size-5">