+++ src/form-elements/FormElementCivicaStreetName.tsx
@@ -100,6 +100,7 @@ function FormElementCivicaStreetName({
           isDirty={isDirty}
           setIsDirty={setIsDirty}
           aria-describedby={ariaDescribedby}
+          autoComplete={element.autocompleteAttributes?.join(' ')}
         />
       </FormElementLabelContainer>