+++ src/form-elements/FormElementCompliance.tsx
@@ -71,8 +71,14 @@ function FormElementCompliance({
       label: 'Notes',
       name: `${element.name}_notes`,
       type: 'textarea',
+      autocompleteAttributes: element.autocompleteAttributes,
     }),
-    [element.id, element.name, element.readOnly],
+    [
+      element.autocompleteAttributes,
+      element.id,
+      element.name,
+      element.readOnly,
+    ],
   )
   const filesElement = React.useMemo<FormTypes.FilesElement>(
     () => ({