+++ src/form-elements/FormElementDateTime.tsx
@@ -112,6 +112,7 @@ function FormElementDateTime({
               disabled={element.readOnly}
               className="input ob-input"
               aria-describedby={ariaDescribedby}
+              autoComplete={element.autocompleteAttributes?.join(' ')}
             />
             <span className="ob-input-icon icon is-small is-right">
               <MaterialIcon className="is-size-5">date_range</MaterialIcon>