+++ src/form-elements/FormElementEmail.tsx
@@ -60,6 +60,7 @@ function FormElementEmail({
               disabled={element.readOnly}
               onBlur={setIsDirty}
               aria-describedby={ariaDescribedby}
+              autoComplete={element.autocompleteAttributes?.join(' ')}
             />
             <span className="ob-input-icon icon is-small is-right">
               <MaterialIcon className="is-size-5">email</MaterialIcon>