+++ src/form-elements/FormElementNumber.tsx
@@ -87,6 +87,7 @@ function FormElementNumber({
                 ref={htmlInputElementRef}
                 onWheel={handleWheel}
                 aria-describedby={ariaDescribedby}
+                autoComplete={element.autocompleteAttributes?.join(' ')}
               />
               <span className="ob-input-icon icon is-small is-right">
                 <MaterialIcon className="is-size-5">tag</MaterialIcon>