+++ src/form-elements/FormElementPointAddress.tsx
@@ -139,6 +139,7 @@ function FormElementPointAddress({
           isDirty={isDirty}
           setIsDirty={setIsDirty}
           aria-describedby={ariaDescribedby}
+          autoComplete={element.autocompleteAttributes?.join(' ')}
         />
       </FormElementLabelContainer>