+++ src/form-elements/FormElementTextarea.tsx
@@ -61,6 +61,7 @@ function FormElementTextarea({
             disabled={element.readOnly}
             onBlur={setIsDirty}
             aria-describedby={ariaDescribedby}
+            autoComplete={element.autocompleteAttributes?.join(' ')}
           />
         </div>